split into python / rust objects
